+++ frontend/packages/core/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./client.js";
+export * from "./errors.js";
 export * from "./oauth.js";
 export * from "./redirects.js";
+export * from "./session.js";